+++ src/main/java/cart/application/IssuableSearchCouponService.java
@@ -1,11 +1,12 @@
 package cart.application;
 
-import cart.domain.Member;
+import cart.domain.member.Member;
 import cart.domain.coupon.Coupon;
 import cart.domain.coupon.CouponRepository;
 import cart.domain.memberCoupon.MemberCoupon;
 import cart.domain.memberCoupon.MemberCouponRepository;
 import cart.dto.IssuableSearchCouponResponse;
+import cart.util.ModelSortHelper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
