+++ src/main/java/cart/domain/cartItem/CartItem.java
@@ -1,5 +1,8 @@
-package cart.domain;
+package cart.domain.cartItem;
 
+import cart.domain.Model;
+import cart.domain.product.Product;
+import cart.domain.member.Member;
 import cart.exception.CartItemException;
 
 import java.util.Objects;
