+++ src/main/java/cart/domain/coupon/CouponUseConditionAction.java
@@ -1,6 +1,6 @@
 package cart.domain.coupon;
 
-import cart.domain.Order;
+import cart.domain.order.Order;
 
 public interface CouponUseConditionAction {
     boolean isUsable(Order order);
