+++ src/main/java/cart/domain/memberCoupon/MemberCoupon.java
@@ -1,6 +1,6 @@
 package cart.domain.memberCoupon;
 
-import cart.domain.Member;
+import cart.domain.member.Member;
 import cart.domain.Model;
 import cart.domain.coupon.Coupon;
 import cart.domain.coupon.DiscountAction;
