+++ src/main/java/cart/dto/ProductResponse.java
@@ -1,6 +1,6 @@
 package cart.dto;
 
-import cart.domain.Product;
+import cart.domain.product.Product;
 
 public class ProductResponse {
     private Long id;
