+++ src/main/java/cart/ui/CartItemApiController.java
@@ -1,8 +1,7 @@
 package cart.ui;
 
 import cart.application.CartItemService;
-import cart.domain.CartItem;
-import cart.domain.Member;
+import cart.domain.member.Member;
 import cart.dto.CartItemQuantityUpdateRequest;
 import cart.dto.CartItemRequest;
 import cart.dto.CartItemResponse;
