+++ src/main/java/cart/ui/MemberCouponController.java
@@ -1,7 +1,7 @@
 package cart.ui;
 
 import cart.application.MemberCouponService;
-import cart.domain.Member;
+import cart.domain.member.Member;
 import cart.dto.MemberCouponRequest;
 import cart.dto.MemberCouponResponse;
 import org.springframework.http.ResponseEntity;
